@@ -1,14 +1,14 @@
 name: random_date
 description: Generate a random date given range. Options let you skip leap years and ignore endYear in that case end value is deduced based on current year plus 5 years.
-version: 0.0.2
+version: 0.0.5
 homepage: https://github.com/NaveenGurram/random-date
 
 environment:
-  sdk: ">=2.8.1 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 #dependencies:
 #  path: ^1.7.0
 
 dev_dependencies:
-  pedantic: ^1.9.0
-  test: ^1.14.4
+  pedantic: ^1.11.1
+  test: ^1.20.1
