@@ -3,19 +3,17 @@ import 'package:test/test.dart';
 
 void main() {
   group('RandomDates', () {
-    RandomDate randomDate;
-    RandomDateOptions randomDateOptions;
+    late RandomDateOptions randomDateOptions;
     setUp(() {
-      randomDate = RandomDate();
       randomDateOptions = RandomDateOptions();
     });
 
     test('justWithStartDate', () {
-      randomDate.generateRandomDate(1980).year > 1980;
+      RandomDate.withStartYear(1980).random().year > 1980;
     });
 
     test('justWithStartDateAndAssertReturnedDate', () {
-      var randomDateTime = randomDate.generateRandomDate(1980);
+      var randomDateTime = RandomDate.withStartYear(1980).random();
       // when end date is not provided, we add +5 to current so date returned
       // has to be less than 2025 and greater than 1980
       var returnedYear = randomDateTime.year;
@@ -23,21 +21,22 @@ void main() {
     });
 
     test('withBothStartAndEndDate', () {
-      var randomDateTime = randomDate.generateRandomDate(1980, 2010);
+      var randomDateTime = RandomDate.withRange(1980, 2010).random();
       randomDateTime.year > 1980 && randomDateTime.year < 2010;
     });
 
     test('withBothDatesAndExcludeLeapYears', () {
-      randomDateOptions.excludeLeapYear = true;
-      var randomDateTime =
-          randomDate.generateRandomDate(2001, 2001, randomDateOptions);
+      var randomDateTime = RandomDate.withRangeAndOptions(
+              2001, 2001, RandomDateOptions.excludeLeapYears())
+          .random();
       randomDateTime.year == 2000;
     });
 
     test('testReturnedYearIsValid', () {
       randomDateOptions.excludeLeapYear = true;
       var randomDateTime =
-          randomDate.generateRandomDate(1981, 2020, randomDateOptions);
+          RandomDate.withRangeAndOptions(1981, 2020, randomDateOptions)
+              .random();
       var output = DateTime.tryParse(randomDateTime.toIso8601String());
       output != null;
     });
@@ -45,20 +44,21 @@ void main() {
     test('withBothDatesAndExcludeLeapYearsWithLeapYearExpectError', () {
       randomDateOptions.excludeLeapYear = true;
       expect(
-          () => randomDate.generateRandomDate(2000, 2000, randomDateOptions),
+          () => RandomDate.withRangeAndOptions(2000, 2000, randomDateOptions)
+              .random(),
           throwsA(allOf(
               isArgumentError,
-              predicate((e) =>
+              predicate((dynamic e) =>
                   e.message ==
                   'Start and End year cannot be same when leap years are excluded'))));
     });
 
     test('withEndYearLessThanStartYear', () {
       expect(
-          () => randomDate.generateRandomDate(2000, 200),
+          () => RandomDate.withRange(2000, 200).random(),
           throwsA(allOf(
               isArgumentError,
-              predicate((e) =>
+              predicate((dynamic e) =>
                   e.message == 'Start year cannot be less then End year'))));
     });
   });
