@@ -1,15 +1,17 @@
 import 'package:random_date/random_date.dart';
 
 void main() {
-  var randomDate = RandomDate();
-  var randomDateOptions = RandomDateOptions();
-  randomDateOptions.addYearsToCurrent = 10;
+  var randomDateOptions = RandomDateOptions.withDefaultYearsToCurrent(10);
   // generates a random date for given range
-  print(
-      'randomDateWithStartYear[2000]: ${randomDate.generateRandomDate(2000)}');
-  print(
-      'randomDateWithRange[2000-2010]: ${randomDate.generateRandomDate(2000, 2010)}');
+  var randomDate = RandomDate.withStartYear(2000);
+  print('randomDateWithStartYear[2000]: ${randomDate.random()}');
+  randomDate = RandomDate.withRange(2000, 2010);
+  print('randomDateWithRange[2000-2010]: ${randomDate.random()}');
   randomDateOptions.excludeLeapYear = true;
+  randomDate = RandomDate.withRangeAndOptions(2000, 2010, randomDateOptions);
   print(
-      'randomDateWithRangeAndSkipLeapYears[2000-2010]: ${randomDate.generateRandomDate(2000, 2010, randomDateOptions)}');
+      'randomDateWithRangeAndSkipLeapYears[2000-2010]: ${randomDate.random()}');
+  randomDate = RandomDate.withStartYearAndOptions(2000, randomDateOptions);
+  print(
+      'randomDateWithStartYearAndEndYearDefaultedTo[2000-2025]: ${randomDate.random()}');
 }
